requirements.txt
anthropic>=0.18.0
paramiko>=3.0.0
Pillow>=10.0.0
